+++ src/cqfantasy/house_wall/__init__.py
@@ -2,3 +2,4 @@ from .tudor_wall import tudor_wall
 from .WallTudor import WallTudor
 from .WallStuccoBrick import WallStuccoBrick
 from .WallTudorPaneling import WallTudorPaneling
+from .WallSplit import WallSplit